Refactored dmarc report detail tables
