+++ api/queries.py
@@ -12,19 +12,19 @@ from schemas.domain import Domain
 from resolvers.domains import resolve_domain, resolve_domains
 
 # Get Dmarc Report Bar Graph Data
-from schemas.dmarc_report_churro_chart import (
+from schemas.dmarc_report_summary_list import (
     get_dmarc_report_churro_chart,
     demo_get_dmarc_report_churro_chart,
 )
 
 # Get Dmarc Report Detail Tables
-from schemas.dmarc_report_detailed_tables import (
-    get_dmarc_report_detailed_tables,
-    get_demo_dmarc_report_detailed_tables,
+from schemas.dmarc_report_detail_tables import (
+    dmarc_report_detail_tables,
+    demo_dmarc_report_detail_tables,
 )
 
 # Get Dmarc Report Doughnut Data
-from schemas.dmarc_report_doughnut import (
+from schemas.dmarc_report_summary import (
     get_dmarc_report_doughnut,
     demo_get_dmarc_report_doughnut,
 )
@@ -191,8 +191,8 @@ class Query(graphene.ObjectType):
     demo_get_dmarc_report_churro_chart = demo_get_dmarc_report_churro_chart
 
     # --- Start Get Dmarc Report Detailed Table
-    get_dmarc_report_detailed_tables = get_dmarc_report_detailed_tables
-    demo_get_dmarc_report_detailed_tables = get_demo_dmarc_report_detailed_tables
+    dmarc_report_detail_tables = dmarc_report_detail_tables
+    demo_dmarc_report_detail_tables = demo_dmarc_report_detail_tables
 
     # --- Start Get Dmarc Report Doughnut ---
     get_dmarc_report_doughnut = get_dmarc_report_doughnut
