+++ api/schemas/dmarc_report_detail_tables/__init__.py
@@ -3,16 +3,16 @@ import graphene
 from enums.period import PeriodEnums
 from scalars.slug import Slug
 from scalars.year import Year
-from schemas.dmarc_report_detailed_tables.dmarc_report_detailed_tables import (
-    DmarcReportDetailedTables,
+from schemas.dmarc_report_detail_tables.dmarc_report_detail_tables import (
+    DmarcReportDetailTables,
 )
-from schemas.dmarc_report_detailed_tables.resolver import (
-    resolve_dmarc_report_detailed_tables,
-    resolve_demo_dmarc_report_detailed_tables,
+from schemas.dmarc_report_detail_tables.resolver import (
+    resolve_dmarc_report_detail_tables,
+    resolve_demo_dmarc_report_detail_tables,
 )
 
-get_dmarc_report_detailed_tables = graphene.Field(
-    lambda: DmarcReportDetailedTables,
+dmarc_report_detail_tables = graphene.Field(
+    lambda: DmarcReportDetailTables,
     domain_slug=graphene.Argument(
         Slug,
         description="The slugified version of the domain you wish to retrieve data for.",
@@ -28,13 +28,13 @@ get_dmarc_report_detailed_tables = graphene.Field(
         description="The year in which the returned data is relevant to.",
         required=True,
     ),
-    resolver=resolve_dmarc_report_detailed_tables,
+    resolver=resolve_dmarc_report_detail_tables,
     description="Query used for gathering data for dmarc report detail tables.",
 )
 
 
-get_demo_dmarc_report_detailed_tables = graphene.Field(
-    lambda: DmarcReportDetailedTables,
+demo_dmarc_report_detail_tables = graphene.Field(
+    lambda: DmarcReportDetailTables,
     domain_slug=graphene.Argument(
         Slug,
         description="The slugified version of the domain you wish to retrieve data for.",
@@ -50,6 +50,6 @@ get_demo_dmarc_report_detailed_tables = graphene.Field(
         description="The year in which the returned data is relevant to.",
         required=True,
     ),
-    resolver=resolve_demo_dmarc_report_detailed_tables,
+    resolver=resolve_demo_dmarc_report_detail_tables,
     description="Query used for gathering data for dmarc report detail tables.",
 )
