+++ api/schemas/dmarc_report_detail_tables/gql_query.py
