+++ api/tests/testdata/get_dmarc_report_detail_table/get_dmarc_report_detail_table_expected_result.py
@@ -1,6 +1,6 @@
 dmarc_report_detail_table_expected_result = {
     "data": {
-        "getDmarcReportDetailedTables": {
+        "dmarcReportDetailTables": {
             "month": "April",
             "year": "2020",
             "detailTables": {
