+++ api/tests/testdata/get_dmarc_report_detail_table/get_dmarc_report_table_query.py
@@ -1,6 +1,6 @@
 test_query = """
 {
-    getDmarcReportDetailedTables(
+    dmarcReportDetailTables(
         domainSlug: "test-domain-gc-ca"
         year: "2020"
         period: MAY
